refactor(widgets): rewrite with public exports
