+++ apps/frontend/src/widgets/footer/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as Footer } from "./Footer.vue";
+export { default as FooterSocials } from "./FooterSocials.vue";
