+++ apps/frontend/src/widgets/header/ui/Header.vue
