+++ apps/frontend/src/widgets/home-hero/ui/Hero.vue
