+++ apps/frontend/src/widgets/home-hero/ui/HeroRecipe.vue
