+++ apps/frontend/src/widgets/home-hero/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as Hero } from "./Hero.vue";
+export { default as HeroRecipe } from "./HeroRecipe.vue";
