+++ apps/frontend/src/widgets/recipe-comments/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as RecipeComments } from "./RecipeComments.vue";
